postgres: collapse into single endpoint file
postgres_runtime.go (wire-protocol gateway) and postgres_auth.go
(SCRAM/cleartext upstream auth) merge into postgres.go alongside the
schema + plugin registration. Same package, all postgres-specific
code in one place — matches the per-provider file convention used
elsewhere in config/plugins/.

No code changes; just file consolidation. Tests + go build clean.
